requirements.txt
foolbox==2.3.0
pytorchcv
pyyaml==3.12
numpy
scipy
pytorch=0.4.1
torchvision
opencv3
yacs
tqdm